+++ src/main/java/com/clinicadental/clinica/repository/IPatientRepository.java
@@ -7,6 +7,6 @@ import org.springframework.stereotype.Repository;
 
 
 @Repository
-public interface IPatientRepository extends JpaRepository<Patient, Long> {
+public interface IPatientRepository extends JpaRepository<Patient, Integer> {
 
 }
